+++ history.md
@@ -1,5 +1,9 @@
 ### History
 
+#### 1.9.0
+- Implement AgeDependentPriorityBlockingQueueTest
+- Implement AgeDependentEntry
+
 #### 1.8.0
 - Implement FixedSizeBlockingQueue, FixedSizeBlockingQueueTest